feat(main/dashboards): separate dashboards routes from dashboards module
